+++ lib/android/app/src/test/java/com/reactnativenavigation/viewcontrollers/NavigatorTest.java
@@ -4,7 +4,6 @@ import android.app.Activity;
 import android.support.annotation.NonNull;
 
 import com.reactnativenavigation.BaseTest;
-import com.reactnativenavigation.layout.Navigator;
 import com.reactnativenavigation.mocks.SimpleViewController;
 import com.reactnativenavigation.mocks.TestStackAnimator;
 import com.reactnativenavigation.utils.CompatUtils;
